[Wip] First set of review changes.
Moved the authentication strategies to their own packages
Added restricted where possible
Use unmodifiable set and drop the unused initializer
