+++ src/test/resources/org/jenkinsci/plugins/casc/integrations/globalmatrixauth/GlobalMatrixStrategy.yml
