+++ src/test/resources/org/jenkinsci/plugins/casc/integrations/rolebasedauth/RoleStrategy1.yml
